+++ Sources/TuistCoreTesting/Utils/MockDeprecator.swift
@@ -1,0 +1,10 @@
+import Foundation
+import TuistCore
+
+public final class MockDeprecator: Deprecating {
+    public var notifyArgs: [(deprecation: String, suggestion: String)] = []
+
+    public func notify(deprecation: String, suggestion: String) {
+        notifyArgs.append((deprecation: deprecation, suggestion: suggestion))
+    }
+}