+++ features/step_definitions/init.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 When(/I initialize a (.+) (.+) named (.+)/) do |platform, product, name|
   system("swift", "run", "tuist", "init", "--path", @dir, "--platform", platform, "--product", product, "--name", name)
 end