+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ### Added
 
-- Up can now be specified via `Setup.swift` https://github.com/tuist/tuist/issues/203 by @marciniwanicki & @kwridan
+- **Breaking** Up can now be specified via `Setup.swift` https://github.com/tuist/tuist/issues/203 by @marciniwanicki & @kwridan
 - Schemes generation https://github.com/tuist/tuist/pull/188 by @pepibumur.
 - Environment variables per target https://github.com/tuist/tuist/pull/189 by @pepibumur.